feat(core): Kanal zum Portal mit Long-Polling, Verteilen und Abgleich
Kanal.abholen() holt Anfragen per Long-Polling, verarbeitet sie über anfragen.bearbeite,
verteilt vor der Antwort geänderte Lesestände und stößt danach die Nachläufe (Mails, PDFs,
Alarme) an. verteilen() schickt alle für das Portal erlaubten Dokumente, deren aktuelle
LesestandVersion höher als die zuletzt gesendete ist – gemerkt in app_setting, unabhängig
davon, ob der 5-Minuten-Job, der Admin-Knopf "Lesestand erzeugen" oder der Kanal selbst
veröffentlicht hat. konto:<id>-Dokumente gehen nur für Kunden mit gesetzter portal_konto_id
raus (in verteilen und abgleichen), sonst blieben Konto-Daten von Vereins-/Abo-/Admin-Kunden
oder gerade gelöschten Konten dauerhaft im Portal liegen. abgleichen() holt die im Portal
bekannten Versionen und schickt fehlende/ältere Dokumente nach. runde()/verteiler_runde()
kapseln Backoff, den Ausfall-/Wiederhergestellt-Alarm nach 30 Minuten und den stündlichen
Abgleich; starte()/stoppe() betreiben das als zwei Threads.

lesestand.publiziere vergibt die Version jetzt als max(bisherige Version + 1, Unixzeit in
Millisekunden), damit ein aus einem Backup wiederhergestelltes Hauptsystem nie eine niedrigere
Version als Portal/Halle vergibt; lesestand_version.version wird dafür zu BIGINT (Migration
0009). main.py verweigert den Start in production bei settings.produktionsfehler und startet
den Kanal im lifespan, wenn ENABLE_KANAL und PORTAL_URL gesetzt sind.

+++ core/beachhub_core/models/system.py
@@ -2,7 +2,7 @@ import uuid
 from datetime import datetime
 from typing import Any
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, String
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -54,7 +54,9 @@ class Audit(UUIDMixin, Base):
 class LesestandVersion(Base):
     __tablename__ = "lesestand_version"
     dokument: Mapped[str] = mapped_column(String(80), primary_key=True)
-    version: Mapped[int] = mapped_column(Integer, default=0, nullable=False)
+    # BigInteger: die Version ist max(bisherige Version + 1, Unixzeit in Millisekunden)
+    # (Ruling Lesestand-Versionen) und übersteigt damit schnell den 32-Bit-Bereich.
+    version: Mapped[int] = mapped_column(BigInteger, default=0, nullable=False)
     signiert_am: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     geaendert: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
